Efficiency + Aesthetic Update
I realized while teaching an intro to WebGL, that there is no need to pass every pixel as a vertex and render it as GL_POINTS, instead I could render a single quad the size of the screen with GL_TRIANGLES, as the work is being done in the fragment shader. Updated the color scheme as well.